misc: Fixes compilation warnings.
The following warnings were noticed on g++ 9.3 on Ubuntu 20.04.

ka/uri/parsing.hpp|1071 col 23| warning: comparison of integer
    expressions of different signedness: ‘unsigned int’ and ‘char’
[-Wsign-compare]
||  1071 |   assert(encoded_char <= std::numeric_limits<char>::max());
||       |          ~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

qi/objectuid.hpp: In instantiation of ‘boost::optional<qi::PtrUid>
    qi::deserializeObjectUid(const R&) [with R = std::basic_string<char>]’:
src/messaging/sessionservice.cpp|217 col 81| required from here
qi/objectuid.hpp|35 col 41| warning: comparison of integer expressions
    of different signedness:
    ‘std::__iterator_traits<__gnu_cxx::__normal_iterator<const char*,
    std::basic_string<char> >, void>::difference_type’ {aka ‘long int’} and
    ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
||    35 |     if (std::distance(begin(r), end(r)) == size(uid))
||       |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~

tests/ka/test_range.cpp|346 col 3| required from here
gtest/gtest.h|1527 col 11| warning: comparison of integer expressions of
    different signedness: ‘const long unsigned int’ and ‘const long int’
    [-Wsign-compare]
||  1527 |   if (lhs == rhs) {
||       |       ~~~~^~~~~~

tests/type/test_property.cpp|399 col 3| required from here
gtest/gtest.h|1527 col 11| warning: ‘qi::Future<T>::operator const
    ValueTypeCast&() const [with T = int; qi::Future<T>::ValueTypeCast =
    int]’ is deprecated: "Use either `then`, `andThen`, `value` or `wait`
    functions instead." [-Wdeprecated-declarations]
||  1527 |   if (lhs == rhs) {
||       |       ~~~~^~~~~~

Reviewed-on: http://gerrit2.aldebaran.lan/1767
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>

+++ ka/uri/parsing.hpp
@@ -1068,7 +1068,7 @@ auto decode_percent_unreserved(std::string const& s) -> std::string {
   char percent = '\0';
   iss >> percent >> std::hex >> encoded_char;
   assert(percent == '%');
-  assert(encoded_char <= std::numeric_limits<char>::max());
+  assert(encoded_char <= static_cast<unsigned int>(std::numeric_limits<char>::max()));
   auto const decoded_char = static_cast<char>(encoded_char);
   // If it is a reserved character, return the percent encoded version.
   if (!grammar::reserved(&decoded_char, &decoded_char + 1).empty()) {
